refactor(a2a): split handler state machines
